Added test and updated AA mutation range.
Added a simple test with known mutation site.
Update common. The simple case I though about that if you have
an AA the rule of pos*3 gives the end of the bp positions, but
they were being used as the beginning. This matched the test I
got from looking at "ARO:3004049" from
https://card.mcmaster.ca/ontology/41114.